pyproject.toml
[tool.poetry]
name = "r-containers"
version = "0.1.0"
description = ""
authors = ["Masatoshi Katabuchi <mattocci27@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^3.0.1"
"ruamel.yaml" = "^0.17.16"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
